requirements.txt
aiosqlite
alembic
asyncpg
fastapi_pagination
python-multipart
httpx
psycopg2-binary
python_dotenv
pytest
pytest_asyncio
pytest_mock
sqlalchemy
sqlalchemy_utils
stackprinter
uvicorn
python-jose[cryptography]
passlib[bcrypt]

pydantic
pydantic_settings